pyproject.toml
[project]
name = "deeproute"
version = "0.1.0"
description = "Multi-layer markdown routing MCP server for agentic code assistants"
readme = "README.md"
authors = [
    { name = "Chris Swanson", email = "chris.swanson@formedics.com" }
]
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]",
    "langchain",
    "langgraph",
    "langchain-anthropic",
    "gitpython",
    "pydantic>=2",
    "anthropic",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[project.scripts]
deeproute = "deeproute.server:main"
